moving functions to utils
